Move getPropertiesMetadata to its own module
